pyproject.toml
[project]
name = "ocs"
version = "0.1.0"
description = "Orion Collective Orchestration System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "sqlalchemy[asyncio]>=2.0.27",
    "aiosqlite>=0.20.0",
    "networkx>=3.2.1",
    "jinja2>=3.1.3",
    "google-genai",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
